+++ packages/firebase_core/firebase_core/ios/firebase_core/Sources/firebase_core/FLTFirebasePlugin.m
@@ -2,7 +2,11 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+#if __has_include("include/firebase_core/FLTFirebasePlugin.h")
 #import "include/firebase_core/FLTFirebasePlugin.h"
+#else
+#import "include/FLTFirebasePlugin.h"
+#endif
 
 // Firebase default app name.
 NSString *_Nonnull const kFIRDefaultAppNameIOS = @"__FIRAPP_DEFAULT";