+++ packages/firebase_core/firebase_core_web/lib/firebase_core_web.dart
@@ -3,8 +3,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-library firebase_core_web;
-
 import 'dart:async';
 import 'dart:js_interop';
 import 'dart:js_interop_unsafe';