+++ packages/firebase_data_connect/firebase_data_connect/example/integration_test/generation_e2e.dart
@@ -30,38 +30,37 @@ void runGenerationTest() {
 
       testWidgets('should have generated correct MutationRef',
           (WidgetTester tester) async {
-        final ref = MoviesConnector.instance.createMovie.ref(
-          genre: 'Action',
-          title: 'The Matrix',
-          releaseYear: 1999,
-          rating: 4.5,
-        );
+        final ref = MoviesConnector.instance
+            .createMovie(
+              genre: 'Action',
+              title: 'The Matrix',
+              releaseYear: 1999,
+            )
+            .rating(4.5);
         expect(ref, isNotNull);
         expect(ref.execute, isNotNull);
       });
 
       testWidgets('should have generated correct QueryRef',
           (WidgetTester tester) async {
-        final ref = MoviesConnector.instance.listMovies.ref();
+        final ref = MoviesConnector.instance.listMovies().ref();
         expect(ref, isNotNull);
         expect(ref.execute, isNotNull);
       });
 
       testWidgets('should have generated correct MutationRef using name',
           (WidgetTester tester) async {
-        final ref = MoviesConnector.instance.addPerson.ref(
-          name: 'Keanu Reeves',
-        );
+        final ref = MoviesConnector.instance.addPerson().name('Keanu Reeves');
         expect(ref, isNotNull);
         expect(ref.execute, isNotNull);
       });
 
       testWidgets('should have generated correct MutationRef using nested id',
           (WidgetTester tester) async {
-        final ref = MoviesConnector.instance.addDirectorToMovie.ref(
-          movieId: 'movieId',
-          personId: AddDirectorToMovieVariablesPersonId(id: 'personId'),
-        );
+        final ref = MoviesConnector.instance
+            .addDirectorToMovie()
+            .movieId('movieId')
+            .personId(AddDirectorToMovieVariablesPersonId(id: 'personId'));
         expect(ref, isNotNull);
         expect(ref.execute, isNotNull);
       });