+++ packages/firebase_data_connect/firebase_data_connect/lib/src/common/common_library.dart
@@ -1,13 +1,22 @@
-// Copyright 2024, the Chromium project authors.  Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
+// Copyright 2024 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 library firebase_data_connect_common;
 
 import 'dart:convert';
 
 import 'package:firebase_app_check/firebase_app_check.dart';
-import 'package:firebase_auth/firebase_auth.dart';
 
 part 'dataconnect_error.dart';
 part 'dataconnect_options.dart';
@@ -49,9 +58,6 @@ abstract class DataConnectTransport {
   /// DataConnect backend configuration.
   DataConnectOptions options;
 
-  /// FirebaseAuth to use to get auth token.
-  FirebaseAuth? auth;
-
   /// FirebaseAppCheck to use to get app check token.
   FirebaseAppCheck? appCheck;
 
@@ -66,12 +72,14 @@ abstract class DataConnectTransport {
       String queryName,
       Deserializer<Data> deserializer,
       Serializer<Variables> serializer,
-      Variables? vars);
+      Variables? vars,
+      String? token);
 
   /// Invokes corresponding mutation endpoint.
   Future<Data> invokeMutation<Data, Variables>(
       String queryName,
       Deserializer<Data> deserializer,
       Serializer<Variables> serializer,
-      Variables? vars);
+      Variables? vars,
+      String? token);
 }