+++ packages/firebase_data_connect/firebase_data_connect/lib/src/network/transport_stub.dart
@@ -1,6 +1,16 @@
-// Copyright 2024, the Chromium project authors.  Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
+// Copyright 2024 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 part of firebase_data_connect_transport;
 
@@ -12,13 +22,11 @@ class TransportStub implements DataConnectTransport {
     this.options,
     this.appId,
     this.sdkType,
-    this.auth,
     this.appCheck,
   );
 
   /// FirebaseAuth
   @override
-  FirebaseAuth? auth;
 
   /// FirebaseAppCheck
   @override
@@ -45,7 +53,8 @@ class TransportStub implements DataConnectTransport {
       String queryName,
       Deserializer<Data> deserializer,
       Serializer<Variables>? serializer,
-      Variables? vars) async {
+      Variables? vars,
+      String? token) async {
     // TODO: implement invokeMutation
     throw UnimplementedError();
   }
@@ -56,7 +65,8 @@ class TransportStub implements DataConnectTransport {
       String queryName,
       Deserializer<Data> deserializer,
       Serializer<Variables>? serialize,
-      Variables? vars) async {
+      Variables? vars,
+      String? token) async {
     // TODO: implement invokeQuery
     throw UnimplementedError();
   }
@@ -67,7 +77,6 @@ DataConnectTransport getTransport(
   DataConnectOptions options,
   String appId,
   CallerSDKType sdkType,
-  FirebaseAuth? auth,
   FirebaseAppCheck? appCheck,
 ) =>
-    TransportStub(transportOptions, options, appId, sdkType, auth, appCheck);
+    TransportStub(transportOptions, options, appId, sdkType, appCheck);