+++ packages/firebase_data_connect/firebase_data_connect/test/src/common/common_library_test.dart
@@ -1,10 +1,21 @@
-// Copyright 2024, the Chromium project authors.  Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
+// Copyright 2024 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 import 'package:firebase_app_check/firebase_app_check.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_data_connect/firebase_data_connect.dart';
+import 'package:firebase_data_connect/src/common/common_library.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 
@@ -94,7 +105,7 @@ void main() {
       final queryName = 'testQuery';
       final deserializer = (json) => json;
       final result = await transport.invokeQuery(
-          queryName, deserializer, emptySerializer, null);
+          queryName, deserializer, emptySerializer, null, null);
 
       expect(result, isNotNull);
     });
@@ -103,7 +114,7 @@ void main() {
       final queryName = 'testMutation';
       final deserializer = (json) => json;
       final result = await transport.invokeMutation(
-          queryName, deserializer, emptySerializer, null);
+          queryName, deserializer, emptySerializer, null, null);
 
       expect(result, isNotNull);
     });
@@ -116,28 +127,27 @@ class TestDataConnectTransport extends DataConnectTransport {
       DataConnectOptions options, String appId, CallerSDKType sdkType,
       {FirebaseAuth? auth, FirebaseAppCheck? appCheck})
       : super(transportOptions, options, appId, sdkType) {
-    this.auth = auth;
     this.appCheck = appCheck;
   }
 
   @override
   Future<Data> invokeQuery<Data, Variables>(
-    String queryName,
-    Deserializer<Data> deserializer,
-    Serializer<Variables>? serializer,
-    Variables? vars,
-  ) async {
+      String queryName,
+      Deserializer<Data> deserializer,
+      Serializer<Variables>? serializer,
+      Variables? vars,
+      String? authToken) async {
     // Simulate query invocation logic here
     return deserializer('{}');
   }
 
   @override
   Future<Data> invokeMutation<Data, Variables>(
-    String queryName,
-    Deserializer<Data> deserializer,
-    Serializer<Variables>? serializer,
-    Variables? vars,
-  ) async {
+      String queryName,
+      Deserializer<Data> deserializer,
+      Serializer<Variables>? serializer,
+      Variables? vars,
+      String? authToken) async {
     // Simulate mutation invocation logic here
     return deserializer('{}');
   }