+++ packages/firebase_data_connect/firebase_data_connect/test/src/core/ref_test.dart
@@ -1,19 +1,51 @@
-// Copyright 2024, the Chromium project authors.  Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
+// Copyright 2024 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 // ignore_for_file: unused_local_variable
 
 import 'dart:async';
+import 'package:http/http.dart' as http;
 
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_data_connect/firebase_data_connect.dart';
+import 'package:firebase_data_connect/src/common/common_library.dart';
+import 'package:firebase_data_connect/src/core/ref.dart';
+import 'package:firebase_data_connect/src/network/rest_library.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
+
+import '../network/rest_transport_test.mocks.dart';
 
 // Mock classes
 class MockDataConnectTransport extends Mock implements DataConnectTransport {}
 
 class MockFirebaseDataConnect extends Mock implements FirebaseDataConnect {}
+
+class DCMockUser extends Mock implements User {
+  int count = 0;
+  List<String?> tokens = ['invalid-token', 'valid-token'];
+  @override
+  Future<String?> getIdToken([bool forceRefresh = false]) {
+    // First return an invalid token, then return a valid token
+    return Future.value(tokens[count++]);
+  }
+}
+
+class MockFirebaseAuth extends Mock implements FirebaseAuth {
+  @override
+  User? get currentUser => DCMockUser();
+}
 
 class MockQueryManager extends Mock implements QueryManager {}
 
@@ -82,6 +114,56 @@ void main() {
       mockDataConnect = MockFirebaseDataConnect();
       serializer = (data) => 'serializedData';
       deserializer = (data) => 'deserializedData';
+    });
+  });
+  group('QueryRef', () {
+    late RestTransport transport;
+    late MockFirebaseDataConnect mockDataConnect;
+    late Serializer<String> serializer;
+    late MockClient mockHttpClient;
+    late Deserializer<String> deserializer;
+
+    setUp(() {
+      mockDataConnect = MockFirebaseDataConnect();
+      when(mockDataConnect.auth).thenReturn(MockFirebaseAuth());
+      mockHttpClient = MockClient();
+      transport = RestTransport(
+        TransportOptions('testhost', 443, true),
+        DataConnectOptions(
+          'testProject',
+          'testLocation',
+          'testConnector',
+          'testService',
+        ),
+        'testAppId',
+        CallerSDKType.core,
+        null,
+      );
+      transport.setHttp(mockHttpClient);
+      mockDataConnect.transport = transport;
+    });
+    test(
+        'query should forceRefresh on ID token if the first request is unauthorized',
+        () async {
+      final mockResponse = http.Response('{"error": "Unauthorized"}', 401);
+      final mockResponseSuccess = http.Response('{"success": true}', 200);
+      final deserializer = (String data) => 'Deserialized Data';
+      int count = 0;
+      QueryRef ref = QueryRef(mockDataConnect, 'operation', transport,
+          deserializer, QueryManager(mockDataConnect), emptySerializer, null);
+
+      when(mockHttpClient.post(any,
+              headers: anyNamed('headers'), body: anyNamed('body')))
+          .thenAnswer((invocation) => [
+                Future.value(mockResponse),
+                Future.value(mockResponseSuccess),
+              ][count++]);
+      final result = await ref.execute();
+
+      expect(result.data, 'Deserialized Data');
+      verify(mockHttpClient.post(any,
+              headers: anyNamed('headers'), body: anyNamed('body')))
+          .called(2);
     });
   });
 }