+++ packages/firebase_messaging/firebase_messaging/ios/Classes/FLTFirebaseMessagingPlugin.h
@@ -9,7 +9,8 @@
 #import <Flutter/Flutter.h>
 #endif
 
-#import <Firebase/Firebase.h>
+@import FirebaseMessaging;
+
 #import <Foundation/Foundation.h>
 #import <UserNotifications/UserNotifications.h>
 #import <firebase_core/FLTFirebasePlugin.h>
@@ -42,6 +43,7 @@ API_AVAILABLE(ios(10.0))
 @interface FLTFirebaseMessagingPlugin : FLTFirebasePlugin <FlutterPlugin,
                                                            FLTFirebasePlugin,
                                                            FIRMessagingDelegate,
+                                                           UIApplicationDelegate,
                                                            UNUserNotificationCenterDelegate>
 #else
 @interface FLTFirebaseMessagingPlugin