+++ packages/firebase_messaging/firebase_messaging/lib/firebase_messaging.dart
@@ -3,8 +3,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-library firebase_messaging;
-
 import 'dart:async';
 
 import 'package:firebase_core/firebase_core.dart';