+++ packages/firebase_ml_model_downloader/firebase_ml_model_downloader/lib/src/firebase_ml_model_downloader.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-part of firebase_ml_model_downloader;
+part of '../firebase_ml_model_downloader.dart';
 
 class FirebaseModelDownloader extends FirebasePluginPlatform {
   FirebaseModelDownloader._({required this.app})