+++ packages/_flutterfire_internals/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.3.46
+
+ - Update a dependency to the latest release.
+
+## 1.3.45
+
+ - Update a dependency to the latest release.
+
+## 1.3.44
+
+ - Update a dependency to the latest release.
+
 ## 1.3.43
 
  - Update a dependency to the latest release.