+++ packages/firebase_app_check/firebase_app_check_platform_interface/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.1.0+40
+
+ - Update a dependency to the latest release.
+
+## 0.1.0+39
+
+ - Update a dependency to the latest release.
+
+## 0.1.0+38
+
+ - Update a dependency to the latest release.
+
 ## 0.1.0+37
 
  - Update a dependency to the latest release.