+++ packages/firebase_auth/firebase_auth_platform_interface/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 7.4.9
+
+ - Update a dependency to the latest release.
+
+## 7.4.8
+
+ - Update a dependency to the latest release.
+
+## 7.4.7
+
+ - Update a dependency to the latest release.
+
 ## 7.4.6
 
  - Update a dependency to the latest release.