+++ packages/firebase_crashlytics/firebase_crashlytics/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 4.1.5
+
+ - Update a dependency to the latest release.
+
+## 4.1.4
+
+ - Update a dependency to the latest release.
+
+## 4.1.3
+
+ - **FIX**(all,apple): use modular headers to import ([#13400](https://github.com/firebase/flutterfire/issues/13400)). ([d7d2d4b9](https://github.com/firebase/flutterfire/commit/d7d2d4b93e7c00226027fffde46699f3d5388a41))
+
 ## 4.1.2
 
  - Update a dependency to the latest release.