+++ packages/firebase_crashlytics/firebase_crashlytics_platform_interface/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 3.6.46
+
+ - Update a dependency to the latest release.
+
+## 3.6.45
+
+ - Update a dependency to the latest release.
+
+## 3.6.44
+
+ - Update a dependency to the latest release.
+
 ## 3.6.43
 
  - Update a dependency to the latest release.