+++ packages/firebase_dynamic_links/firebase_dynamic_links_platform_interface/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.2.6+46
+
+ - Update a dependency to the latest release.
+
+## 0.2.6+45
+
+ - Update a dependency to the latest release.
+
+## 0.2.6+44
+
+ - Update a dependency to the latest release.
+
 ## 0.2.6+43
 
  - Update a dependency to the latest release.