+++ packages/firebase_performance/firebase_performance_platform_interface/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.1.4+46
+
+ - Update a dependency to the latest release.
+
+## 0.1.4+45
+
+ - Update a dependency to the latest release.
+
+## 0.1.4+44
+
+ - Update a dependency to the latest release.
+
 ## 0.1.4+43
 
  - Update a dependency to the latest release.