+++ packages/firebase_performance/firebase_performance_web/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.1.7+4
+
+ - Update a dependency to the latest release.
+
+## 0.1.7+3
+
+ - Update a dependency to the latest release.
+
+## 0.1.7+2
+
+ - Update a dependency to the latest release.
+
 ## 0.1.7+1
 
  - Update a dependency to the latest release.